+++ src/storage-adapter-local/src/storage.js
@@ -1,0 +1,10 @@
+import { Storage as StorageCore } from '@brighter/storage'
+
+const Storage = config => {
+  return StorageCore({
+    ...config,
+    type: 'local'
+  })
+}
+
+export { Storage }