+++ src/storage-adapter-local/test/integration.js
@@ -1,0 +1,15 @@
+import chai from 'chai'
+import sinon from 'sinon'
+import sinonChai from 'sinon-chai'
+
+chai.use(sinonChai)
+
+before(() => {
+  global.expect = chai.expect
+  global.sinon = sinon
+})
+
+after(() => {
+  delete global.sinon
+  delete global.expect
+})